new file:   recipq.tex
Nicholas:

"Peeter, I’m trying to understand going from (2.92) to (2.93).  It seems like you simply divide by the same scalar that you divided by in order to make the unit values.  Am I right?  Is this a general approach?  Is there some more meaningful insight to be gained here?  How does (2.92) imply (2.93)?
When the curvilinear basis is orthogonal, as is the case here, then computing the reciprocal basis is as simple as rescaling, as done in (2.93).  If the curvilinear basis vectors are not orthogonal, you have to do a more general computation (for 3D, see (2.21) and for 2D see (2.17))

And to rescale, you find the scaling that yields the unit vectors, and then scale by that squared.  Am I right?"